refactor: modificando retorno null

+++ src/main/java/baozi/store/baozi/controller/ClienteController.java
@@ -6,7 +6,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/cliente")
@@ -24,23 +26,42 @@ public class ClienteController {
     }
 
     @GetMapping
-    public ResponseEntity<List<Cliente>> GetAllCliente() {
+    public ResponseEntity<List<Cliente>> getAllCliente() {
         return ResponseEntity.ok(clienteService.getAll());
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Cliente> GetById(@PathVariable Long id) {
+    public ResponseEntity<?> getById(@PathVariable Long id) {
         Cliente cliente = clienteService.getById(id);
 
         if (cliente == null) {
-            ResponseEntity.notFound().build();
+            Map<String, String> error = new HashMap<>();
+            error.put("message", "Cliente não encontrado.");
+
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND)
+                    .body(error);
         }
-        return ResponseEntity.ok(clienteService.getById(id));
+
+        return ResponseEntity.ok(cliente);
     }
 
     @DeleteMapping("/{id}")
-    public void DeleteById(@PathVariable Long id) {
+    public ResponseEntity<?> deleteById(@PathVariable Long id) {
+        Cliente cliente = clienteService.getById(id);
+
+        if (cliente == null) {
+            Map<String, String> error = new HashMap<>();
+            error.put("message", "Cliente não encontrado.");
+
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND)
+                    .body(error);
+        }
+
         clienteService.deleteCliente(id);
+
+        return ResponseEntity.ok().build();
     }
 
 }
